automation/requirements.txt
# 보탬e 자동화 프로젝트 의존성
# Python 3.9+

# 브라우저 자동화
playwright==1.40.0

# 설정 관리
pyyaml==6.0.1
python-dotenv==1.0.0

# 로깅
loguru==0.7.2

# HTTP 클라이언트
httpx==0.25.2

# 데이터 처리
pandas==2.1.3

# 스케줄링
schedule==1.2.1

# 알림
slack-sdk==3.23.0

# 테스트
pytest==7.4.3
pytest-asyncio==0.21.1

# 타입 체크
mypy==1.7.1
